+++ ext/geos_c_impl/extconf.rb
@@ -45,6 +45,7 @@ if have_header("geos_c.h")
   have_func("GEOSPreparedDisjoint_r", "geos_c.h")
   have_func("GEOSUnaryUnion_r", "geos_c.h")
   have_func("GEOSCoordSeq_isCCW_r", "geos_c.h")
+  have_func("GEOSDensify", "geos_c.h")
   have_func("rb_memhash", "ruby.h")
   have_func("rb_gc_mark_movable", "ruby.h")
 end