+++ test/geos_capi/polygon_test.rb
@@ -179,4 +179,17 @@ class GeosPolygonTest < Minitest::Test # :nodoc:
 
     assert_equal expected, input.polygonize
   end
+
+  def test_segmentize
+    skip_geos_version_less_then("3.10")
+
+    input = @factory.parse_wkt("POLYGON ((0 0, 10 0, 10 10, 0 10, 0 0))")
+    expected = @factory.parse_wkt("POLYGON ((0 0, 5 0, 10 0, 10 5, 10 10, 5 10, 0 10, 0 5, 0 0))")
+
+    assert_equal expected, input.segmentize(5)
+
+    assert_raises(TypeError, "no implicit conversion to float from string") { input.segmentize("a") }
+    assert_raises(TypeError, "no implicit conversion to float from nil") { input.segmentize(nil) }
+    assert_raises(RGeo::Error::InvalidGeometry, "Tolerance must be positive") { input.segmentize(0) }
+  end
 end