+++ History.md
@@ -6,6 +6,7 @@
 * Fix Naming/VariableNumber from rubocop_todos. (@haroon26) #351
 * Fix Lint/MissingSuper from rubocop_todos. (@haroon26) #352
 * Drop support for Ruby 2.6. (@seuros) #353
+* Add `segmentize` method to CAPI geometires (@arkirchner) #364
 
 **Bug Fixes**